platform/polyglot/static_app_loader/py/pyproject.toml
[project]
name = "static-app-loader"
version = "1.0.0"
description = "Polyglot static app loader for FastAPI with SPA routing, template engines, and multi-app support"
license = { text = "MIT" }
requires-python = ">=3.11,<3.12"
keywords = ["fastapi", "static", "spa", "polyglot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "fastapi==0.128.0",
    "pydantic==2.12.5",
    "jinja2>=3.1.0",
    "python-liquid>=1.0.0",
    "aiofiles==25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx[http2]>=0.28.1",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "uvicorn[standard]==0.40.0",
]

[project.scripts]
static-app-loader = "static_app_loader.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/static_app_loader"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["__tests__"]
